Creation of dataset V1
This class aims to create the dataset necessary for the AI to use. It gets information from the grid and from it collect 6 different datas: teh peaks, highest,lowest and average height, the bumpiness and number of holes. Fianlly the get_data method aims to collect all of this data and return it